+++ app/Http/Routes/V2/UserRoute.php
@@ -10,7 +10,7 @@ class UserRoute
     {
         $router->group([
             'prefix' => 'user',
-            'middleware' => ['user.frontend', 'user']
+            'middleware' => 'user'
         ], function ($router) {
             // User
             $router->get('/resetSecurity', [UserController::class, 'resetSecurity']);